fix: bump .zenodo.json version 0.9.4 → 0.9.5
Pre-commit version consistency hook caught this miss. Now 0.9.5 in all
6 locations: pyproject.toml, __init__.py, CITATION.cff, .zenodo.json,
RESULTS_SUMMARY.md, CHANGELOG.md.

+++ src/scpn_quantum_control/gauge/vortex_detector.py
@@ -95,7 +95,7 @@ def plaquette_vorticity(phases: np.ndarray, plaquette: list[int]) -> int:
         i = plaquette[idx]
         j = plaquette[(idx + 1) % len(plaquette)]
         total += _angle_diff(float(phases[i]), float(phases[j]))
-    return round(total / (2 * np.pi))
+    return int(round(total / (2 * np.pi)))
 
 
 def measure_vortex_density(
